requirements.txt
allure-python-commons
playwright
pytest~=8.4.0
requests~=2.32.3
faker~=37.3.0
pydantic~=2.11.7
python-dotenv~=1.1.1